Started Working on Question Layout.
Zoomed in the whole view for easy reading. Created a partial to use as a
template for questions.

Styled question partial to be like the header (perhaps I'm overdoing the
flag motif) Still, it looks kind of nice and having a common theme will
likely tie everything together well.

Added form/buttons to the question for 'listening' and 'being heard'.
Aka, looking at what others say and what you've said. Maybe I should
change it to say "Your opinion" and "Other opinions"

Anyway, I intend on using demographic data to visualize the information
so people can see what parts of the world agree with them and what dont.
I think that will be a useful and interesting measure. Also, I've made
the current statistic be 'politics' category. I want to eventually have
lots of categories of questions, but I'm focussing on simple for now.

Thought: Use Harp's _data.json, but have a batch job or something update
the _data.json with the top questions and etcera every few hours or so.
Then use a service/ajax to have more dynamic content.