Added Javascript Loading System.
Javascript files for specific files can be loaded via duplicating their
file path in javascript (so /test/index.html would load /test/index.js)
if the javascript file exists.

Also, working on a partial loading system, right now it's just setting a
flag on the global variable which might not be the best practice in the
word.

Created script.ejs partial to load scripts and also created a couple of
globals:

		"excludeJQuery" : false,
		"noJS" : false,
		"javascriptPath" : "/js",

so that we can shut off js entirely if desired, and also set the base
path to look for javascript files.