Added statistics page to be ajaxed.
Going to use colorbox or something similar to pull in the statistics
page for whatever question is asked.

Created some dummy data, we're going to need to return the states to the
statistics page some time and load it into the map correctly later, but
for now we define the map's properties in the dummy data for global.

Still need to wire in click events to the map, but this is a good
committing point so that I can pull in color box and get that working
with the other people's opinions button.