Began Work on tutorial elements.
Created css file and partial for the tutorial.

Updated index page to include the partial, note that we must include the
css from within the scripts partial, we cannot include it in the head
otherwise it would never be included since at that point we have no
inkling as to whether or not the files to support the tutorial should be
included or not.