Question info and cursors.
Also updated header to have a sub heading somewhat. I wonder if the 'On'
is neccesary in the navigation. Probably not.

Added cursors for the more info span and a crosshair on the loaded map
statistics. I still need to use a world map, get high charts in there
and all that good stuff. Ah well. Such is life.