Renamed statistics, added color and help button
Changed categories

Added help button to top left

Added kular colors for green as a base. not sure about it yet, but
because css var's it will be easy to change.

Renamed statistics to info becuase it will be used to get information
about the question / load up an interaction page with it.