Added AJAX form submission and slide up.
With that in place, really the only thing left to deal with is
HighCharts and making the statistics/info page have a layout that uses
both the world map (with drilldown) and information about whatever the
question is.

After that is done then it will be time to decide the API to be used
(perhaps an experiment in GoLang?) and to begin constructing the service
achitecture. And once that is all done the prototype should be ready and
I can perhaps enlist @scperkins to help me play with pure or another
framework to make the site look cleaner.