Updated settings for geoip.
Set GEO path as well as contrib module for gis.

Next thing to do (ServerSide) is to figure out how best to convert ip
information from opinions/questions into data that can be easily shown
upon a map.

THe most likely scenario to take would be to store ip and the resolved
address during the creation of the opinion and then use the aggregate of
that information for display. The big question is, how will we display
it? Clearly we have the source of information (the ip), but what would
be most helpful in storing for mapping purposes?

For a USA Map clearly we'd want to use state, and for a country sized
map we'd want to use countries. But what about other countries? I'm
unfamilar with mapping there, region? subregion? zones?